backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
requests==2.31.0
reportlab==4.1.0
gunicorn==21.2.0
python-docx==1.1.0
pdfplumber==0.11.0
PyMuPDF==1.24.2
python-docx
reportlab
textract
pypandoc
